chore(quality): raise PHPStan to level 7 — 56 → 0 errors
Tighten type contracts across the codebase so static analysis at
level 7 reports no issues. Changes are correctness improvements,
not silent widenings:

- SplFileObject::current() may return string|array|false; guard with
  is_string() before trim/rtrim (admin-debug.php, logger.php).
- wp_json_encode() returns string|false; handle the false case when
  building cache keys (audit-trail.php).
- strtotime() returns int|false; fall back to time() before gmdate()
  (audit-trail.php, export-stats.php, logger-enhanced.php).
- realpath() returns string|false; cast for filesize()/readfile()
  (batch-file-handler.php).
- glob() returns array|false; check is_array() before merging
  (diagnostics.php).
- WP_Hook is iterable in WP core; cast to array before iterating
  (diagnostics.php).
- preg_split() returns list<string>|false; guard before foreach
  (helpers.php, markdown-exporter.php).
- wp_remote_retrieve_body()/wp_remote_retrieve_header() may yield
  array; coerce to string explicitly (image-processor.php).
- imagecreatetruecolor() needs int<1,max>; clamp new_height to >=1.
- imagecolorallocatealpha() returns int|false; guard imagefill().
- ->get_results() rows are stdClass; cast to array before
  property access (logger-enhanced.php, session.php).
- get_post() returns WP_Post|array|null; use instanceof WP_Post
  narrowing (export-query-controller.php, page-collector.php).
- get_term() returns WP_Term|WP_Error|false; instanceof WP_Term
  narrowing (seo-reader.php).
- get_bloginfo() returns string|false; coerce before substr()
  (page-collector.php).
- get_posts() returns array<WP_Post>|false; guard is_array()
  (page-collector.php).
- preg_match() alternatives yield possibly-undefined captures;
  default via isset() (pdf-exporter.php).
- validate_export_result(): tighten object → SScribe_Result so
  is_success()/get_data()/get_error() resolve (batch-processor.php).
- max() on possibly-empty array; widen to is_array + non-empty
  guard (docx-content-renderer.php).
- normalize_ip_literal(): long2ip() returns string|false; map to
  string|null per declared return type (exporter.php).
- Export rate limiter: cast (int) max() result so decision carries
  int  (export-rate-limiter.php).
- random_bytes(): guard  with max(1, )
  (batch-processor.php).

phpstan.neon: level 6 → 7.

Regression: PHPUnit 941/941 passing (21 intentionally skipped),
PHPCS clean, ESLint clean.

+++ includes/class-sscribe-export-rate-limiter.php
@@ -187,7 +187,7 @@ class SScribe_Export_Rate_Limiter {
 		return SScribe_Rate_Limit_Decision::allowed(
 			$bucket,
 			$rate_limit,
-			max( 0, $rate_limit - $new_count ),
+			(int) max( 0, $rate_limit - $new_count ),
 			(int) $data['reset_at']
 		);
 	}
